@@ -4,7 +4,7 @@
     "homepage": "https://auth0.com/wordpress",
     "license": "GPLv2",
     "require": {
-        "php": "^5.3 || ^7.0"
+        "php": "^5.6 || ^7.0"
     },
     "require-dev": {
         "dealerdirect/phpcodesniffer-composer-installer": "^0.4.3",
