@@ -31,14 +31,14 @@ class JWT {
 	 */
 	public static $timestamp = null;
 
-	public static $supported_algs = array(
-		'HS256' => array( 'hash_hmac', 'SHA256' ),
-		'HS512' => array( 'hash_hmac', 'SHA512' ),
-		'HS384' => array( 'hash_hmac', 'SHA384' ),
-		'RS256' => array( 'openssl', 'SHA256' ),
-		'RS384' => array( 'openssl', 'SHA384' ),
-		'RS512' => array( 'openssl', 'SHA512' ),
-	);
+	public static $supported_algs = [
+		'HS256' => [ 'hash_hmac', 'SHA256' ],
+		'HS512' => [ 'hash_hmac', 'SHA512' ],
+		'HS384' => [ 'hash_hmac', 'SHA384' ],
+		'RS256' => [ 'openssl', 'SHA256' ],
+		'RS384' => [ 'openssl', 'SHA384' ],
+		'RS512' => [ 'openssl', 'SHA512' ],
+	];
 
 	/**
 	 * Decodes a JWT string into a PHP object.
@@ -60,7 +60,7 @@ class JWT {
 	 * @uses jsonDecode
 	 * @uses urlsafeB64Decode
 	 */
-	public static function decode( $jwt, $key, array $allowed_algs = array() ) {
+	public static function decode( $jwt, $key, array $allowed_algs = [] ) {
 		$timestamp = is_null( static::$timestamp ) ? time() : static::$timestamp;
 
 		if ( empty( $key ) ) {
@@ -147,17 +147,17 @@ class JWT {
 	 * @uses urlsafeB64Encode
 	 */
 	public static function encode( $payload, $key, $alg = 'HS256', $keyId = null, $head = null ) {
-		$header = array(
+		$header = [
 			'typ' => 'JWT',
 			'alg' => $alg,
-		);
+		];
 		if ( $keyId !== null ) {
 			$header['kid'] = $keyId;
 		}
 		if ( isset( $head ) && is_array( $head ) ) {
 			$header = array_merge( $head, $header );
 		}
-		$segments      = array();
+		$segments      = [];
 		$segments[]    = static::urlsafeB64Encode( static::jsonEncode( $header ) );
 		$segments[]    = static::urlsafeB64Encode( static::jsonEncode( $payload ) );
 		$signing_input = implode( '.', $segments );
@@ -338,12 +338,12 @@ class JWT {
 	 * @return void
 	 */
 	private static function handleJsonError( $errno ) {
-		$messages = array(
+		$messages = [
 			JSON_ERROR_DEPTH          => 'Maximum stack depth exceeded',
 			JSON_ERROR_STATE_MISMATCH => 'Invalid or malformed JSON',
 			JSON_ERROR_CTRL_CHAR      => 'Unexpected control character found',
 			JSON_ERROR_SYNTAX         => 'Syntax error, malformed JSON',
-		);
+		];
 		throw new DomainException(
 			isset( $messages[ $errno ] )
 			? $messages[ $errno ]
