@@ -3,22 +3,22 @@ ini_set( 'display_errors', 1 );
 require_once( 'TwitterAPIExchange.php' );
 
 /** Set access tokens here - see: https://dev.twitter.com/apps/ */
-$settings = array(
+$settings = [
 	'oauth_access_token'        => '',
 	'oauth_access_token_secret' => '',
 	'consumer_key'              => '',
 	'consumer_secret'           => '',
-);
+];
 
 /** URL for REST request, see: https://dev.twitter.com/docs/api/1.1/ */
 $url           = 'https://api.twitter.com/1.1/blocks/create.json';
 $requestMethod = 'POST';
 
 /** POST fields required by the URL above. See relevant docs as above */
-$postfields = array(
+$postfields = [
 	'screen_name' => 'usernameToBlock',
 	'skip_status' => '1',
-);
+];
 
 /** Perform a POST request and echo the response */
 $twitter = new TwitterAPIExchange( $settings );
