@@ -1,5 +1,5 @@
 <?php
-function renderAuth0Form( $canShowLegacyLogin = true, $specialSettings = array() ) {
+function renderAuth0Form( $canShowLegacyLogin = true, $specialSettings = [] ) {
 	if ( is_user_logged_in() ) {
 		return;
 	}
@@ -16,13 +16,13 @@ function renderAuth0Form( $canShowLegacyLogin = true, $specialSettings = array()
 			wp_enqueue_script( 'wpa0_auth0js', $auth0_js_url, false, null, true );
 		}
 
-		wp_enqueue_script( 'wpa0_lock', $options->get_lock_url(), array( 'jquery' ), false, true );
+		wp_enqueue_script( 'wpa0_lock', $options->get_lock_url(), [ 'jquery' ], false, true );
 		wp_enqueue_script( 'js-cookie', WPA0_PLUGIN_LIB_URL . 'js.cookie.min.js', false, '2.2.0', true );
-		wp_enqueue_script( 'wpa0_lock_init', WPA0_PLUGIN_JS_URL . 'lock-init.js', array( 'jquery' ), WPA0_VERSION, true );
+		wp_enqueue_script( 'wpa0_lock_init', WPA0_PLUGIN_JS_URL . 'lock-init.js', [ 'jquery' ], WPA0_VERSION, true );
 		wp_localize_script(
 			'wpa0_lock_init',
 			WP_Auth0_Lock10_Options::LOCK_GLOBAL_JS_VAR_NAME,
-			array(
+			[
 				'settings'        => $lock_options->get_lock_options(),
 				'ready'           => WP_Auth0::ready(),
 				'domain'          => $lock_options->get_domain(),
@@ -33,20 +33,20 @@ function renderAuth0Form( $canShowLegacyLogin = true, $specialSettings = array()
 				'loginFormId'     => WPA0_AUTH0_LOGIN_FORM_ID,
 				'showAsModal'     => ! empty( $specialSettings['show_as_modal'] ),
 				'ssoOpts'         => $check_sso ? $lock_options->get_sso_options() : null,
-				'i18n'            => array(
+				'i18n'            => [
 					'notReadyText'       => __( 'Auth0 is not configured', 'wp-auth0' ),
 					'cannotFindNodeText' => __( 'Auth0 cannot find node with id ', 'wp-auth0' ),
 					'modalButtonText'    => ! empty( $specialSettings['modal_trigger_name'] )
 					  ? sanitize_text_field( $specialSettings['modal_trigger_name'] )
 					  : __( 'Login', 'wp-auth0' ),
-				),
-			)
+				],
+			]
 		);
 
 		$login_tpl = apply_filters( 'auth0_login_form_tpl', 'auth0-login-form.php', $lock_options, $canShowLegacyLogin );
 		require $login_tpl;
 	} else {
-		add_action( 'login_footer', array( 'WP_Auth0', 'render_back_to_auth0' ) );
-		add_action( 'woocommerce_after_customer_login_form', array( 'WP_Auth0', 'render_back_to_auth0' ) );
+		add_action( 'login_footer', [ 'WP_Auth0', 'render_back_to_auth0' ] );
+		add_action( 'woocommerce_after_customer_login_form', [ 'WP_Auth0', 'render_back_to_auth0' ] );
 	}
 }
