@@ -59,7 +59,7 @@ class WP_Auth0_EditProfile {
 	 * @codeCoverageIgnore - Deprecated.
 	 */
 	public function init() {
-		add_action( 'admin_enqueue_scripts', array( $this, 'admin_enqueue_scripts' ) );
+		add_action( 'admin_enqueue_scripts', [ $this, 'admin_enqueue_scripts' ] );
 	}
 
 	/**
@@ -71,14 +71,14 @@ class WP_Auth0_EditProfile {
 	public function admin_enqueue_scripts() {
 		global $pagenow;
 
-		if ( ! in_array( $pagenow, array( 'profile.php', 'user-edit.php' ) ) ) {
+		if ( ! in_array( $pagenow, [ 'profile.php', 'user-edit.php' ] ) ) {
 			return;
 		}
 
 		wp_enqueue_script(
 			'wpa0_user_profile',
 			WPA0_PLUGIN_JS_URL . 'edit-user-profile.js',
-			array( 'jquery' ),
+			[ 'jquery' ],
 			WPA0_VERSION
 		);
 
@@ -88,19 +88,19 @@ class WP_Auth0_EditProfile {
 		wp_localize_script(
 			'wpa0_user_profile',
 			'wpa0UserProfile',
-			array(
+			[
 				'userId'         => intval( $GLOBALS['user_id'] ),
 				'userStrategy'   => sanitize_text_field( $strategy ),
 				'deleteIdNonce'  => wp_create_nonce( 'delete_auth0_identity' ),
 				'deleteMfaNonce' => wp_create_nonce( 'delete_auth0_mfa' ),
 				'ajaxUrl'        => admin_url( 'admin-ajax.php' ),
-				'i18n'           => array(
+				'i18n'           => [
 					'confirmDeleteId'   => __( 'Are you sure you want to delete the Auth0 user data for this user?', 'wp-auth0' ),
 					'actionComplete'    => __( 'Deleted', 'wp-auth0' ),
 					'actionFailed'      => __( 'Action failed, please see the Auth0 error log for details.', 'wp-auth0' ),
 					'cannotChangeEmail' => __( 'Email cannot be changed for non-database connections.', 'wp-auth0' ),
-				),
-			)
+				],
+			]
 		);
 	}
 
